pyproject.toml
[tool.poetry]
name = "dashboard"
version = "0.1.0"
description = ""
authors = ["jhpark <jhpark@chowis.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
psycopg2 = "^2.9.5"
pillow = "^9.2.0"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
